backend/requirements.txt
aiosqlite==0.19.0
alembic==1.12.0
asyncpg==0.28.0
fastapi==0.103.1
pydantic==2.3.0
pydantic-settings==2.0.3
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
SQLAlchemy==2.0.21
uvicorn==0.23.2
redis==5.0.0
prometheus-client==0.17.1
structlog==23.1.0
pyotp==2.9.0
passlib[bcrypt]==1.7.4
psycopg2-binary==2.9.7
python-dotenv==1.0.0
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0
pytest-cov==4.1.0
aioredis==2.0.1
psutil==5.9.5